# api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Documentation",
      "version": "1.0.0"
    },
    "paths": {
      "/api/packets": {
        "get": {
          "tags": ["Packets"],
          "summary": "Returns a list of packets",
          "responses": {
            "200": {
              "description": "A JSON array of packet objects",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PacketDetail"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
            "tags": ["Packets"],
            "summary": "Creates a new packet",
            "requestBody": {
              "required": true,
              "content": {
                "multipart/form-data": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "files": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "binary"
                        }
                      },
                      "data": {
                        "type": "string",
                        "description": "Stringified JSON with the packet details"
                      }
                    }
                  },
                  "encoding": {
                    "document": {
                      "contentType": "application/pdf",
                      "style": "form"
                    },
                    "data": {
                      "contentType": "application/json"
                    }
                  }
                }
              }
            },
            "responses": {
              "201": {"description": "Packet created successfully"},
              "400": {"description": "Invalid input data"}
            }
          }
        },
     
      "/api/packets/{packetId}": {
        "get": {
          "tags": ["Packets"],
          "summary": "Gets a packet by ID",
          "parameters": [
            {
              "in": "path",
              "name": "packetId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The packet's ID"
            }
          ],
          "responses": {
            "200": {
              "description": "A packet object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PacketDetail"
                  }
                }
              }
            },
            "404": {"description": "Packet not found"}
          }
        },
        "put": {
          "tags": ["Packets"],
          "summary": "Edits a packet by ID",
          "parameters": [
            {
              "in": "path",
              "name": "packetId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The packet's ID"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacketDetail"
                }
              }
            }
          },
          "responses": {
            "200": {"description": "Packet updated successfully"},
            "404": {"description": "Packet not found"}
          }
        },
        "delete": {
          "tags": ["Packets"],
          "summary": "Deletes a packet by ID",
          "parameters": [
            {
              "in": "path",
              "name": "packetId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The packet's ID"
            }
          ],
          "responses": {
            "200": {"description": "Packet deleted successfully"},
            "404": {"description": "Packet not found"}
          }
        }
      },
      "/api/packets/{packetId}/files/{fileId}": {
        "get": {
          "tags": ["Packets"],
          "summary": "Downloads a file by ID",
          "parameters": [
            {
              "in": "path",
              "name": "packetId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The packet's ID"
            },
            {
              "in": "path",
              "name": "fileId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The file's ID"
            }
          ],
          "responses": {
            "200": {
              "description": "File downloaded successfully",
              "content": {
                "application/octet-stream": {
                  "schema": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "404": {"description": "File not found"}
          }
        },
        "delete": {
          "tags": ["Packets"],
          "summary": "Deletes a file by ID",
          "parameters": [
            {
              "in": "path",
              "name": "packetId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The packet's ID"
            },
            {
              "in": "path",
              "name": "fileId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The file's ID"
            }
          ],
          "responses": {
            "200": {"description": "File deleted successfully"},
            "404": {"description": "File not found"}
          }
        }
      },
      "/api/users": {
        "get": {
          "tags": ["Users"],
          "summary": "Returns a list of users",
          "responses": {
            "200": {
              "description": "A JSON array of user objects",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/users/{userId}": {
        "get": {
          "tags": ["Users"],
          "summary": "Gets a user by ID",
          "parameters": [
            {
              "in": "path",
              "name": "userId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The user's ID"
            }
          ],
          "responses": {
            "200": {
              "description": "A user object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {"description": "User not found"}
          }
        }
      },
      "/api/users/{userId}/packets": {
        "get": {
          "tags": ["Users"],
          "summary": "Returns a list of packets associated with a user",
          "responses": {
            "200": {
              "description": "A JSON array of packet objects",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PacketDetail"
                    }
                  }
                }
              }
            }
          }
        
        }, "post": {
            "tags": ["Users"],
            "summary": "Creates a new packet",
            "requestBody": {
              "required": true,
              "content": {
                "multipart/form-data": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "files": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "binary"
                        }
                      },
                      "data": {
                        "type": "string",
                        "description": "Stringified JSON with the packet details"
                      }
                    }
                  },
                  "encoding": {
                    "document": {
                      "contentType": "application/pdf",
                      "style": "form"
                    },
                    "data": {
                      "contentType": "application/json"
                    }
                  }
                }
              }
            },
            "responses": {
              "201": {"description": "Packet created successfully"},
              "400": {"description": "Invalid input data"}
            }
          }
        
      },
      "/api/users/{userId}/packets/{packetId}/phases": {
        "get": {
          "tags": ["Packets"],
          "summary": "Returns a list of phases associated with a packet",
          "responses": {
            "200": {
              "description": "A JSON array of packet phase objects",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PacketPhase"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Packets"],
          "summary": "Creates a new packet phase",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PacketPhase"
                  }
              }
            }
          }
          },
          "responses": {
            "201": {"description": "Packet phase created successfully"},
            "400": {"description": "Invalid input data"}
          }
        },
        "put": {
          "tags": ["Packets"],
          "summary": "Updates an existing packet phase",
          "parameters": [
            {
              "in": "path",
              "name": "phaseId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The ID of the packet phase"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacketPhase"
                }
              }
            }
          },
          "responses": {
            "200": {"description": "Packet phase updated successfully"},
            "400": {"description": "Invalid input data"},
            "404": {"description": "Packet phase not found"}
          }
        }
      },
      "/api/users/{userId}/packets/{packetId}/files/{fileId}": {
        "get": {
          "tags": ["Files"],
          "summary": "Download a PDF file",
          "description": "Downloads a specific PDF file associated with a user's packet.",
          "parameters": [
            {
              "in": "path",
              "name": "userId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The ID of the user"
            },
            {
              "in": "path",
              "name": "packetId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The ID of the packet"
            },
            {
              "in": "path",
              "name": "fileId",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The ID of the file to download"
            }
          ],
          "responses": {
            "200": {
              "description": "The PDF file",
              "content": {
                "application/pdf": {
                  "schema": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "404": {"description": "File not found"},
            "401": {"description": "Unauthorized - User does not have permission to access this file"}
          }
        }
      }
    },
    "components": {
      "schemas": {
        "PacketDetail": {
          "type": "object",
          "properties": {
            "packetId": {"type": "string"},
            "name": {"type": "string"},
            "type": {"type": "string"},
            "comments": {"type": "string", "nullable": true},
            "currentPhase": {"type": "integer"},
            "creator": {"type": "string"},
            "createdAt": {"type": "string", "format": "date-time"},
            "files": {
              "type": "array",
              "items": {"$ref": "#/components/schemas/File"}
            },
            "creatorUser": {"$ref": "#/components/schemas/User"},
            "phases": {
              "type": "array",
              "items": {"$ref": "#/components/schemas/PacketPhase"}
            }
          }
        },
        "File": {
          "type": "object",
          "properties": {
            "fileId": {"type": "string"},
            "name": {"type": "string"},
            "createdAt": {"type": "string", "format": "date-time"},
            "packetId": {"type": "string"}
          }
        },
        "User": {
          "type": "object",
          "properties": {
            "userId": {"type": "string"},
            "firstName": {"type": "string"},
            "lastName": {"type": "string"},
            "grade": {"type": "string"},
            "rank": {"type": "string"},
            "branch": {"type": "string"},
            "organizationId": {"type": "string"},
            "raterId": {"type": "string", "nullable": true},
            "email": {"type": "string"}
          }
        },
     
        "PacketPhase": {
          "type": "object",
          "properties": {
            "packetPhaseId": {"type": "string"},
            "suspense": {"type": "string", "format": "date"},
            "comments": {"type": "string", "nullable": true},
            "stepNumber": {"type": "integer"},
            "completionDate": {"type": "string", "format": "date-time", "nullable": true},
            "packetId": {"type": "string"},
            "phase": {"type": "string"},
            "assignee": {"type": "string"},
            "assigneeUser": {"$ref": "#/components/schemas/User"}
          }
        }
      },
      "responses": {
        "PacketCreated": {
          "description": "Packet created successfully"
        },
        "InvalidInput": {
          "description": "Invalid input data"
        },
        "FileDownloaded": {
          "description": "File downloaded successfully"
        },
        "FileNotFound": {
          "description": "File not found"
        },
        "PacketUpdated": {
          "description": "Packet updated successfully"
        },
        "PacketDeleted": {
          "description": "Packet deleted successfully"
        },
        "UserNotFound": {
          "description": "User not found"
        },
        "UnauthorizedFileAccess": {
          "description": "Unauthorized - User does not have permission to access this file"
        }
      }
    
    }
  }
  